finish

+++ controllers/contactsController.js
@@ -1,10 +1,16 @@
 const { Contact } = require("../models");
 
 const getContacts = async (req, res, next) => {
-  console.log("GET /api/contacts called");
-
   try {
-    const data = await Contact.find({});
+    const { _id: owner } = req.user;
+    const { page = 1, limit = 20, favorite } = req.query;
+    const query = favorite ? { owner, favorite } : { owner };
+
+    const skip = (page - 1) * limit;
+    const data = await Contact.find(query, "", {
+      skip,
+      limit: Number(limit),
+    }).populate("owner", "_id email");
     console.log(data);
     res.status(200).json(data);
   } catch (error) {
@@ -15,7 +21,8 @@ const getContacts = async (req, res, next) => {
 const getContactnbyId = async (req, res, next) => {
   try {
     const { contactId } = req.params;
-    const data = await Contact.findById({ _id: contactId });
+    const { _id: userId } = req.user;
+    const data = await Contact.findOne({ _id: contactId, owner: userId });
     if (!data) {
       return res.status(404).json({ message: "Not found" });
     }
@@ -26,10 +33,10 @@ const getContactnbyId = async (req, res, next) => {
 };
 
 const postContact = async (req, res, next) => {
-  console.log("POST /api/contacts directed...");
-  console.log(req.body);
+  const { _id: owner } = req.user;
+
   try {
-    const data = await Contact.create(req.body);
+    const data = await Contact.create({ ...req.body, owner });
     return res.status(201).json(data);
   } catch (error) {
     next(error);
@@ -39,11 +46,16 @@ const postContact = async (req, res, next) => {
 const deleteContact = async (req, res, next) => {
   try {
     const { contactId } = req.params;
-    const data = await Contact.findByIdAndDelete(contactId);
-    if (data) {
-      return res.status(200).json({ message: "contact deleted" });
+    const { _id: userId } = req.user;
+
+    const contact = await Contact.findOne({ _id: contactId, owner: userId });
+    if (!contact) {
+      return res
+        .status(404)
+        .json({ message: "Not found or not owned by user" });
     }
-    res.status(400).json({ message: "Not found" });
+    await Contact.findByIdAndDelete(contactId);
+    return res.status(200).json({ message: "contact deleted" });
   } catch (error) {
     next(error);
   }
@@ -52,40 +64,54 @@ const deleteContact = async (req, res, next) => {
 const putContact = async (req, res, next) => {
   try {
     const { contactId } = req.params;
+    const { _id: userId } = req.user;
 
-    const data = await Contact.findByIdAndUpdate(contactId, req.body, {
-      new: true,
-    });
-    if (!data) {
-      return res.status(404).json({ message: "Not found" });
+    const contact = await Contact.findOne({ _id: contactId, owner: userId });
+    if (!contact) {
+      return res
+        .status(404)
+        .json({ message: "Not found or not owned by user" });
     }
-    return res.status(200).json(data);
+
+    const updatedContact = await Contact.findByIdAndUpdate(
+      contactId,
+      req.body,
+      { new: true }
+    );
+    return res.status(200).json(updatedContact);
   } catch (erorr) {
     next(erorr);
   }
 };
 
- const  updateStatusContact = async (contactId, updateBody) => 
- {
-    return await Contact.findByIdAndUpdate(contactId, updateBody, { new: true });
-}
-
+const updateStatusContact = async (userId, contactId, updateBody) => {
+  const contact = await Contact.findOne({ _id: contactId, owner: userId });
+  if (!contact) {
+    throw new Error("Not found or not owned by user");
+  }
+  const updatedContact = await Contact.findByIdAndUpdate(
+    contactId,
+    updateBody,
+    { new: true }
+  );
+  return updatedContact;
+};
 
 const updateContactFavorite = async (req, res, next) => {
-
   try {
     const { contactId } = req.params;
     const { favorite } = req.body;
 
     if (favorite === undefined) {
       return res.status(400).json({ message: "missing field favorite" });
-  }
+    }
 
-  const updatedContact = await updateStatusContact(contactId, { favorite });
+    const userId = req.user._id; 
+    const updatedContact = await updateStatusContact(userId, contactId, { favorite });
 
-  if (!updatedContact) {
-    return res.status(404).json({ message: "Not found"});
-}
+    if (!updatedContact) {
+      return res.status(404).json({ message: "Not found" });
+    }
 
     return res.status(200).json(updatedContact);
   } catch (error) {
